Migrate _auth logic into the ReqServerChannel
